+++ src/main/java/az/spring/request/AccountRequest.java
@@ -4,13 +4,15 @@ import az.spring.entity.User;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 public class AccountRequest {
 
     private String accountNumber;
     private Boolean status;
-    private Double balance;
+    private BigDecimal balance;
     private User user;
 
 }
